adding rizzino_a09b.js
adding rizzino_a09b.js.  All parts should be working, however I am unsure if the display of the element with the ID attribute value of b001 is displaying what is desired for this assignment.  Please let me know if you have a suggestion of how to display this correctly if I have not done so already. I have already tried several extensions to the variable to display the output of what the 2 paragraphs in this div display in the html file but I am having trouble finding the one that will display the text correctly.